### 2017-06-02 #120 @kaneplusplus
Submission for R Package bittrex
Package bittrex is an R implementation of the REST interface used by the Bittrex crypto-currency exchange. It provides functions for endpoints supported by the exchange. This includes the ability to retrieve price, volume, and order book information as well as the ability to trade crypto-currencies.

```
Package: bittrex
Title: Client for the Bitrex Exchange
Version: 0.1.0
Authors@R: c(person("Michael", "Kane", role = c("aut", "cph", "cre"), email = "kaneplusplus@gmail.com"))
Description: A client for the Bitrex crypo-currency exchange \url{https://bittrex.com/} including the ability to query trade data, manage account balances, and place orders.
License: LGPL-2
Imports: httr, openssl
Remotes: bwlewis/rthreejs
Suggests: covr, testthat, knitr, rmarkdown, threejs, tidyverse, igraph, quantmod
```

https://github.com/kaneplusplus/bittrex

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under and why? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

The target audience consists of finance researchers and traders interested in crypo-currencies.

No package provide an interface to the [Bittrex cryptocurrency exchange](https://c-cex.com/) meant for research, which includes the ability to trade. Package [Rbitcoin](https://github.com/jangorecki/Rbitcoin) is an "end-to-end trading engine in R" for currency trading with Bitcoin. It is not clear of other crypto-currencies are included. Package [IBroker](https://github.com/joshuaulrich/IBrokers) is an API for the [Interactive Broker](https://www.interactivebrokers.com/en/home.php) trading platform. It provides the ability to trade stocks and currencies, not crypto-currencies.

Kyle Hamilton: @kylehamilton
Taylor Arnold: @statsmaths
