### 2017-06-04 #120 @noamross
### Editor checks:

---

#### Editor comments

Thanks, @kaneplusplus.

---

Reviewers:
Due date:
