### 2017-06-18 #120 @statsmaths
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is a very well-written client for placing API calls to Bitrex crypo-currency exchange. There is excellent coverage of both unit tests and documentation; the README file got me up and running in under 5 minutes. I have only two issues and one suggestion before publishing the package:

- Include a maintainer entry in the DESCRIPTION file.

- add a link to the CONDUCT.md from the README.md file; if I understand correctly, this is the preferred method from the packaging_guide.md file.

- (suggestion) I had previously reviewed the **ccex** package, which ran into trouble because the API was throttled and eventually shut-down. It would be nice if this package had something along the lines of a `bittrex_api_check` function that returns `TRUE` if the API seems to be working and `FALSE` otherwise. This would make tracking down bug-fixes down the line easier and would be a helpful function to use when writing a script that makes use of the package.
