### 2017-06-19 #120 @kaneplusplus
WRT the maintainer entry: With apologies for quoting documentation... according to [Writing R Extensions](https://cran.r-project.org/doc/manuals/R-exts.html#The-DESCRIPTION-file) "The Package’, ‘Version’, ‘License’, ‘Description’, ‘Title’, ‘Author’, and ‘Maintainer’ fields are mandatory, all other fields are optional. Fields ‘Author’ and ‘Maintainer’ can be auto-generated from ‘Authors@R’, and may be omitted if the latter is provided." If the latter is provided then the "cre" role is designated as maintainer. I'm happy to add it though, to be more explicit.

WRT link to CONDUCT.md: Thanks for pointing this out. The link has been added.

WRT `bittrex_api_check`: This is a great suggesion and it has been added to the package. The function returns `TRUE` when the `GET` request's status code is 200 and `FALSE` otherwise. By default a warning is provided with the status code if the request is not successful.
