### 2017-06-20 #120 @statsmaths
@kaneplusplus Thanks for the quick turn around on the response / fixes. The new `bittrex_api_check` is exactly what I had in mind. I'll leave the whole implicit/explicit Maintainer field up to the editor, but otherwise this addresses all of my concerns.
