### 2017-06-24 #120 @noamross
Package Review
--------------

_(As I haven't heard from a second reviewer I went ahead and did an editor's review)_

-   [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

> ##### Paper (for packages co-submitting to JOSS)
>
> The package contains a `paper.md` with:
>
> -   [x] **A short summary** describing the high-level functionality of the software
> -   [x] **Authors:** A list of authors with their affiliations
> -   [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> -   [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is an excellent package providing access to data for Bittrex currency market research and interaction. It has an intuitive R API, covers the Bittrex REST API completely, is easy to set up and follow best practices of REST API interaction.

-   I appreciate the type-stable/error-safe responses of the API calling functions. I would try to extend this in a few places. or `getorderbook()`, I would suggest returning a single data frame with a `type` column with `buy` and `sell` values. (Essentially the lines in the vignette). Also, empty values (e.g., from `getopenorders()`, the `get*histotory()` functions) are returned as a list rather than a zero-row data frame. I would suggest returning the latter with the expected fields.
-   Running `getbalance("something")` when I have no currency of type "something" yields an error. Can this fail more gracefully?

-   For `getticker()`, I suggest adding to the docs that the market names are the names column from `getmarkets()`, and they are not case-sensitive.
-   In `getorderbook()` doc - can you specify what `depth` means? I assumed it would be number of records but the rows are much larger than this. Also, I suggest using an example with a market that yields results. `usd-btc` yields `INVALID_MARKET`.
-   Something is left out in the `getmarkethistory` `market` parameter description.
-   When you note that "market orders a currently disabled", can you specify whether this disabled in your package or in the Bittrex API? It might also be worth it to put a date here if it's the bittrex API because this could stop being true. (e.g., "As of DATE/VERSION...")
-   I am submitting a PR with some typo corrections, including what I believe were places you accidentally wrote "C-cex" instead of "Bittrex"
-   For private interfaces, it would be helpful to show results in the examples of what the output should look like for successful calls / transactions. Running the examples just for testing will yield failures, as they should, so example successful output (commented out in the not-run example), will be instructive.

#### Other Suggestions

I consider none of these mandatory but please consider them.

-   It would he helpful to as a function prefix (such at `bt_`) to API call functions. API function calls are fairly similar to what one might expect from other financial API packages (e.g., `buy`), and this helps avoid namespace clashes and enables discovery through autocompletion. For programming people use `::` but for interactive use this is very helpful.
-   You might consider returning the output data frames as tibbles, so that if the user has `tidyverse` packages loaded the shorter and more informative print methods are used, but if not they will act as standard data frames.
-   POSIXct times are also more `tibble()` compatible. Unless POSIXlt is required for compatibility with likely used packages like *xts* or *quantmod* I suggest POSIXct as the default time storage format.
-   I would suggest breaking up the `bittrex.r` file into multiple files for future ease of maintainability/collaboration.

Per @statsmaths' comment, the automated generation of `Maintainer:` field is fine with us.  I'll make a note to clarify this in our guidelines.
