### 2017-07-11 #120 @kaneplusplus
@noamross Thanks very much for the pull request, comments, and suggestions. I have made all of the revisions except for one. I decided not to include the `tibble` class information despite liking the approach you suggested. It gives slight preference to the tidyverse over `data.table`. However, I have added an issue, as an enhancement, the ability to register a function that `data.frame`s generated in the package would be passed to. This would allow users to decorate returned `data.frame`s in whichever way they want. Is this sufficient?

If so, then I think the package is ready to go. Thanks again for the review. It was extremely helpful.
