### 2017-07-11 #120 @noamross
Looks good! Tiny things to fix and then we'll be good to go:
-  for `bt_getorderbook()`, the example reflects the previous format of output (list rather than data frame)
-  Please use `<-` instead of `=` for assignment. That was in the `goodpractice()` output but I neglected to call it out before.
-  The vignette still uses the old function names.
-  I fixed what I think is a documentation typo in a pull request.

The `tibble()` idea was just a suggestion.  That's fine. I commented on the issue.
