### 2017-07-12 #120 @kaneplusplus
Sorry, I should have caught these. The revisions have been made.
