### 2017-07-12 #120 @noamross
Approved! Thanks @kaneplusplus for submitting and @statsmaths for your review!

To-dos:
- [ ] Transfer the repo to the rOpenSci organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.
```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
```
- [ ] Add the (brand new!) "Peer-reviewed" badge to your README (in the future people will add this on submittal; it updates through the stages of review):
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (we'll turn on the services on our end.)
- [ ] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). Let me know if you are interested.
