### 2017-07-14 #120 @kaneplusplus
@noamross I'm having some trouble with Zenodo. I'm able to log in and see most of my repositories, but I'm not able to see the bittrex repository. Did I need to register the package before transferring the repo to the organization?
