### 2017-07-14 #120 @kaneplusplus
That did it. Thanks. I'll have the rest of the checklist in the next hour or so.

I am interested in a blog post. However, I probably won't be able to get to this until next week.
