### 2017-07-14 #120 @kaneplusplus
OK, badges have been added, Zenodo is watching, a DOI has been created, and I have submitted to JOSS.

Is the blog in one of the repositories? If not, please tell me who it should be sent to?
